refactor(fft): address PR review feedback for Bowers FFT
Address reviewer comments from PR #288:

Safety improvements:
- Add bounds checks with descriptive panic messages to `row()`, `row_mut()`
- Add bounds check to `get_layer()` method
- Use `checked_mul` for row index calculation to prevent overflow
- Add `debug_assert!` for twiddle array length validation
- Add overflow protection with `1usize.checked_shl()` in `LayerTwiddles::new`

Code organization:
- Move all tests to separate `bowers_fft_tests.rs` file
- Add `num_layers()` method to `LayerTwiddles` for better encapsulation
- Reduce main implementation file from 988 to 693 lines

Documentation:
- Update IFFT docs to clarify it doesn't scale by 1/n
- Add panic documentation to bounds-checked methods
- Remove verbose inline comments (moved rationale to docs)

All 22 tests pass including new bounds-check tests.
