+++ crypto/math/src/fft/cpu/bowers_fft_tests.rs
@@ -1,0 +1,376 @@
+//! Tests for Bowers FFT implementation
+//!
+//! Separated from the main implementation to keep the library code concise.
+
+use super::bowers_fft::*;
+use crate::fft::cpu::bit_reversing::in_place_bit_reverse_permute;
+use crate::fft::cpu::fft::in_place_nr_2radix_fft;
+use crate::fft::cpu::roots_of_unity::get_twiddles;
+use crate::field::element::FieldElement;
+use crate::field::fields::fft_friendly::u64_goldilocks::GoldilocksField;
+use crate::field::traits::{IsFFTField, RootsConfig};
+use alloc::vec;
+use alloc::vec::Vec;
+use proptest::{collection, prelude::*};
+
+type F = GoldilocksField;
+type FE = FieldElement<F>;
+
+/// Naive O(n²) DFT for correctness verification
+pub fn naive_dft(input: &[FE]) -> Vec<FE> {
+    let n = input.len();
+    if n == 0 {
+        return Vec::new();
+    }
+    let root = F::get_primitive_root_of_unity(n.trailing_zeros() as u64).unwrap();
+    let mut result = vec![FE::zero(); n];
+
+    for (k, res) in result.iter_mut().enumerate() {
+        for (j, inp) in input.iter().enumerate() {
+            *res = &*res + &(inp * &root.pow((j * k) as u64));
+        }
+    }
+
+    result
+}
+
+// =========================================================================
+// FftMatrix tests
+// =========================================================================
+
+#[test]
+fn test_fft_matrix_roundtrip() {
+    let polys = vec![
+        (0..8).map(|i| FE::from(i as u64)).collect::<Vec<_>>(),
+        (8..16).map(|i| FE::from(i as u64)).collect::<Vec<_>>(),
+    ];
+
+    let matrix = FftMatrix::from_polynomials(polys.clone());
+    assert_eq!(matrix.width, 8);
+    assert_eq!(matrix.height, 2);
+
+    let recovered = matrix.to_polynomials();
+    assert_eq!(recovered, polys);
+}
+
+#[test]
+fn test_fft_matrix_empty() {
+    let polys: Vec<Vec<FE>> = vec![];
+    let matrix = FftMatrix::from_polynomials(polys);
+    assert_eq!(matrix.width, 0);
+    assert_eq!(matrix.height, 0);
+}
+
+#[test]
+#[should_panic(expected = "Row index out of bounds")]
+fn test_fft_matrix_row_out_of_bounds() {
+    let polys = vec![(0..8).map(|i| FE::from(i as u64)).collect::<Vec<_>>()];
+    let matrix = FftMatrix::from_polynomials(polys);
+    let _ = matrix.row(1); // Should panic
+}
+
+#[test]
+#[should_panic(expected = "Row index out of bounds")]
+fn test_fft_matrix_row_mut_out_of_bounds() {
+    let polys = vec![(0..8).map(|i| FE::from(i as u64)).collect::<Vec<_>>()];
+    let mut matrix = FftMatrix::from_polynomials(polys);
+    let _ = matrix.row_mut(1); // Should panic
+}
+
+// =========================================================================
+// LayerTwiddles tests
+// =========================================================================
+
+#[test]
+fn test_layer_twiddles_creation() {
+    let order = 4u64;
+    let layer_twiddles = LayerTwiddles::<F>::new(order).unwrap();
+
+    assert_eq!(layer_twiddles.layers.len(), 4);
+    assert_eq!(layer_twiddles.layers[0].len(), 8);
+    assert_eq!(layer_twiddles.layers[1].len(), 4);
+    assert_eq!(layer_twiddles.layers[2].len(), 2);
+    assert_eq!(layer_twiddles.layers[3].len(), 1);
+
+    // First twiddle of each layer should be 1
+    for layer in &layer_twiddles.layers {
+        assert_eq!(layer[0], FE::one());
+    }
+}
+
+#[test]
+fn test_layer_twiddles_overflow_protection() {
+    // Order 64 would overflow on 64-bit systems
+    let result = LayerTwiddles::<F>::new(64);
+    assert!(result.is_none());
+}
+
+#[test]
+#[should_panic(expected = "Layer index out of bounds")]
+fn test_layer_twiddles_get_layer_out_of_bounds() {
+    let layer_twiddles = LayerTwiddles::<F>::new(4).unwrap();
+    let _ = layer_twiddles.get_layer(10); // Should panic
+}
+
+// =========================================================================
+// Bowers FFT tests
+// =========================================================================
+
+#[test]
+fn test_bowers_fft_opt_fused_small() {
+    let input: Vec<FE> = (0..4).map(|i| FE::from(i as u64)).collect();
+    let expected = naive_dft(&input);
+
+    let layer_twiddles = LayerTwiddles::<F>::new(2).unwrap();
+    let mut result = input.clone();
+    bowers_fft_opt_fused(&mut result, &layer_twiddles).unwrap();
+    in_place_bit_reverse_permute(&mut result);
+
+    assert_eq!(result, expected);
+}
+
+#[test]
+fn test_bowers_fft_opt_fused_medium() {
+    let input: Vec<FE> = (0..16).map(|i| FE::from(i as u64)).collect();
+    let expected = naive_dft(&input);
+
+    let layer_twiddles = LayerTwiddles::<F>::new(4).unwrap();
+    let mut result = input.clone();
+    bowers_fft_opt_fused(&mut result, &layer_twiddles).unwrap();
+    in_place_bit_reverse_permute(&mut result);
+
+    assert_eq!(result, expected);
+}
+
+#[test]
+fn test_bowers_fft_opt_fused_large() {
+    let input: Vec<FE> = (0..256).map(|i| FE::from(i as u64)).collect();
+    let expected = naive_dft(&input);
+
+    let layer_twiddles = LayerTwiddles::<F>::new(8).unwrap();
+    let mut result = input.clone();
+    bowers_fft_opt_fused(&mut result, &layer_twiddles).unwrap();
+    in_place_bit_reverse_permute(&mut result);
+
+    assert_eq!(result, expected);
+}
+
+#[test]
+fn test_bowers_fft_size_one() {
+    let input: Vec<FE> = vec![FE::from(42u64)];
+    let layer_twiddles = LayerTwiddles::<F>::new(0).unwrap();
+    let mut result = input.clone();
+    bowers_fft_opt_fused(&mut result, &layer_twiddles).unwrap();
+    assert_eq!(result, input);
+}
+
+#[test]
+fn test_bowers_fft_size_two() {
+    let input: Vec<FE> = vec![FE::from(1u64), FE::from(2u64)];
+    let expected = naive_dft(&input);
+
+    let layer_twiddles = LayerTwiddles::<F>::new(1).unwrap();
+    let mut result = input.clone();
+    bowers_fft_opt_fused(&mut result, &layer_twiddles).unwrap();
+    in_place_bit_reverse_permute(&mut result);
+
+    assert_eq!(result, expected);
+}
+
+#[test]
+fn test_bowers_batch_fft_opt() {
+    let polys = vec![
+        (0..8).map(|i| FE::from(i as u64)).collect::<Vec<_>>(),
+        (8..16).map(|i| FE::from(i as u64)).collect::<Vec<_>>(),
+    ];
+
+    let expected: Vec<Vec<FE>> = polys.iter().map(|p| naive_dft(p)).collect();
+
+    let mut matrix = FftMatrix::from_polynomials(polys);
+    let layer_twiddles = LayerTwiddles::<F>::new(3).unwrap();
+    bowers_batch_fft_opt(&mut matrix, &layer_twiddles).unwrap();
+
+    let result = matrix.to_polynomials();
+    assert_eq!(result, expected);
+}
+
+#[test]
+fn test_bowers_fft_non_power_of_two() {
+    use crate::fft::errors::FFTError;
+    let input: Vec<FE> = (0..7).map(|i| FE::from(i as u64)).collect();
+    let layer_twiddles = LayerTwiddles::<F>::new(3).unwrap();
+    let mut result = input;
+    let err = bowers_fft_opt_fused(&mut result, &layer_twiddles);
+    assert!(matches!(err, Err(FFTError::InputError(7))));
+}
+
+// =========================================================================
+// Bowers vs Native FFT comparison tests
+// =========================================================================
+
+fn compare_bowers_vs_native(input: &[FE]) {
+    let order = input.len().trailing_zeros() as u64;
+
+    // Native Cooley-Tukey FFT
+    let native_twiddles = get_twiddles(order, RootsConfig::BitReverse).unwrap();
+    let mut native_result = input.to_vec();
+    in_place_nr_2radix_fft::<F, F>(&mut native_result, &native_twiddles);
+    in_place_bit_reverse_permute(&mut native_result);
+
+    // Bowers FFT
+    let layer_twiddles = LayerTwiddles::<F>::new(order).unwrap();
+    let mut bowers_result = input.to_vec();
+    bowers_fft_opt_fused(&mut bowers_result, &layer_twiddles).unwrap();
+    in_place_bit_reverse_permute(&mut bowers_result);
+
+    assert_eq!(bowers_result, native_result);
+}
+
+#[test]
+fn test_bowers_vs_native_various_sizes() {
+    for order in 0..=10u64 {
+        let input: Vec<FE> = (0..(1 << order)).map(|i| FE::from(i as u64)).collect();
+        compare_bowers_vs_native(&input);
+    }
+}
+
+#[test]
+fn test_bowers_vs_native_edge_cases() {
+    // All zeros
+    let zeros: Vec<FE> = (0..64).map(|_| FE::zero()).collect();
+    compare_bowers_vs_native(&zeros);
+
+    // All ones
+    let ones: Vec<FE> = (0..64).map(|_| FE::one()).collect();
+    compare_bowers_vs_native(&ones);
+
+    // Alternating
+    let alternating: Vec<FE> = (0..64)
+        .map(|i| if i % 2 == 0 { FE::zero() } else { FE::one() })
+        .collect();
+    compare_bowers_vs_native(&alternating);
+}
+
+// =========================================================================
+// Property-based tests (proptest)
+// =========================================================================
+
+prop_compose! {
+    fn field_element()(num in any::<u64>()) -> FE {
+        FE::from(num)
+    }
+}
+
+prop_compose! {
+    fn field_vec(max_exp: u8)(exp in 1u8..=max_exp)(
+        vec in collection::vec(field_element(), 1usize << exp)
+    ) -> Vec<FE> {
+        vec
+    }
+}
+
+proptest! {
+    #![proptest_config(ProptestConfig::with_cases(50))]
+
+    #[test]
+    fn proptest_bowers_matches_naive_dft(coeffs in field_vec(8)) {
+        let expected = naive_dft(&coeffs);
+
+        let order = coeffs.len().trailing_zeros() as u64;
+        let layer_twiddles = LayerTwiddles::<F>::new(order).unwrap();
+
+        let mut result = coeffs;
+        bowers_fft_opt_fused(&mut result, &layer_twiddles).unwrap();
+        in_place_bit_reverse_permute(&mut result);
+
+        prop_assert_eq!(result, expected);
+    }
+
+    #[test]
+    fn proptest_bowers_matches_native_fft(coeffs in field_vec(8)) {
+        let order = coeffs.len().trailing_zeros() as u64;
+
+        // Native FFT
+        let native_twiddles = get_twiddles(order, RootsConfig::BitReverse).unwrap();
+        let mut native_result = coeffs.clone();
+        in_place_nr_2radix_fft::<F, F>(&mut native_result, &native_twiddles);
+        in_place_bit_reverse_permute(&mut native_result);
+
+        // Bowers FFT
+        let layer_twiddles = LayerTwiddles::<F>::new(order).unwrap();
+        let mut bowers_result = coeffs;
+        bowers_fft_opt_fused(&mut bowers_result, &layer_twiddles).unwrap();
+        in_place_bit_reverse_permute(&mut bowers_result);
+
+        prop_assert_eq!(bowers_result, native_result);
+    }
+}
+
+// =========================================================================
+// Parallel tests
+// =========================================================================
+
+#[cfg(feature = "parallel")]
+mod parallel_tests {
+    use super::*;
+
+    #[test]
+    fn test_bowers_fft_opt_fused_parallel_small() {
+        // Small input - should use sequential path
+        let input: Vec<FE> = (0..16).map(|i| FE::from(i as u64)).collect();
+        let expected = naive_dft(&input);
+
+        let layer_twiddles = LayerTwiddles::<F>::new(4).unwrap();
+        let mut result = input.clone();
+        bowers_fft_opt_fused_parallel(&mut result, &layer_twiddles).unwrap();
+        in_place_bit_reverse_permute(&mut result);
+
+        assert_eq!(result, expected);
+    }
+
+    #[test]
+    fn test_bowers_fft_opt_fused_parallel_medium() {
+        // Medium input
+        let input: Vec<FE> = (0..256).map(|i| FE::from(i as u64)).collect();
+        let expected = naive_dft(&input);
+
+        let layer_twiddles = LayerTwiddles::<F>::new(8).unwrap();
+        let mut result = input.clone();
+        bowers_fft_opt_fused_parallel(&mut result, &layer_twiddles).unwrap();
+        in_place_bit_reverse_permute(&mut result);
+
+        assert_eq!(result, expected);
+    }
+
+    #[test]
+    fn test_bowers_fft_opt_fused_parallel_large() {
+        // Large input - should exercise parallel paths
+        let input: Vec<FE> = (0..4096).map(|i| FE::from(i as u64)).collect();
+        let expected = naive_dft(&input);
+
+        let layer_twiddles = LayerTwiddles::<F>::new(12).unwrap();
+        let mut result = input.clone();
+        bowers_fft_opt_fused_parallel(&mut result, &layer_twiddles).unwrap();
+        in_place_bit_reverse_permute(&mut result);
+
+        assert_eq!(result, expected);
+    }
+
+    #[test]
+    fn test_parallel_matches_sequential() {
+        // Verify parallel and sequential produce identical results
+        let input: Vec<FE> = (0..1024).map(|i| FE::from(i as u64)).collect();
+
+        let layer_twiddles = LayerTwiddles::<F>::new(10).unwrap();
+
+        let mut result_seq = input.clone();
+        bowers_fft_opt_fused(&mut result_seq, &layer_twiddles).unwrap();
+        in_place_bit_reverse_permute(&mut result_seq);
+
+        let mut result_par = input.clone();
+        bowers_fft_opt_fused_parallel(&mut result_par, &layer_twiddles).unwrap();
+        in_place_bit_reverse_permute(&mut result_par);
+
+        assert_eq!(result_seq, result_par);
+    }
+}
