+++ crypto/math/src/fft/cpu/mod.rs
@@ -1,6 +1,8 @@
 pub mod bit_reversing;
 #[cfg(feature = "alloc")]
 pub mod bowers_fft;
+#[cfg(all(test, feature = "alloc"))]
+mod bowers_fft_tests;
 pub mod fft;
 #[cfg(feature = "alloc")]
 pub mod ops;
